holiday_management/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.0.2
django-cors-headers==4.3.1
django-query-counter==0.4.0
djangorestframework==3.14.0
idna==3.10
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2025.1
requests==2.31.0
six==1.17.0
sqlparse==0.5.3
tabulate==0.9.0
tzdata==2025.1
urllib3==2.3.0
